requirements.txt
apache-airflow[statsd]